Ensure code key is set on the data dictionary
There are cases where new codes may be added to the API and they will be added to the info_message dictionary; however, there is a case where the 'code' key is not set at all. This modification checks that the 'code' key is at least set and then checks if it exists in the info_message and codes dictionary.